Move ControllerType and RedstoneMode to common package.
