--- refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/network/ControllerType.java
@@ -1,6 +1,0 @@
-package com.refinedmods.refinedstorage2.platform.api.network;
-
-public enum ControllerType {
-    NORMAL,
-    CREATIVE
-}
