+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/block/ControllerType.java
@@ -1,0 +1,6 @@
+package com.refinedmods.refinedstorage2.platform.common.block;
+
+public enum ControllerType {
+    NORMAL,
+    CREATIVE
+}
