+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/block/entity/RedstoneModeSettings.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage2.platform.common.block.entity;
 
-import com.refinedmods.refinedstorage2.platform.api.network.node.RedstoneMode;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.TwoWaySyncProperty;
+import com.refinedmods.refinedstorage2.platform.common.util.RedstoneMode;
 
 public class RedstoneModeSettings {
     private static final int IGNORE = 0;
